backend/requirements.txt
fastapi
uvicorn
pillow
python-multipart
transformers
timm
# Torch and torchvision are heavy; install them if you plan to run real inference.
# Recommended (CPU):
# torch==2.2.0+cpu torchvision==0.18.1+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
# Or install via: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
